backend/skillNexus/requirements.txt
django==4.2.0
djangorestframework
pyyaml
requests
django-cors-headers
djangorestframework-simplejwt
Pillow
django-rest-swagger
drf-yasg
drf-spectacular
C:\xampp\htdocs\backend\env\Lib\site-packages\rest_framework_swagger\templates\swagger_ui.html
C:\xampp\htdocs\backend\env\Lib\site-packages\rest_framework_swagger\templates\swagger-ui.html